server/pyproject.toml
[project]
name = "readme-server"
version = "0.1.0"
description = "Kid Reading Voice App – server (API + bot)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "loguru>=0.7.3",
    "pipecat-ai[daily,webrtc,deepgram,google,hume,local-smart-turn-v3,runner,silero]>=0.0.102",
    "pipecatcloud>=0.2.6",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.0",
    "aiortc>=1.14.0",
    "dramatiq[redis]>=1.18.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
    "ty>=0.0.1",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
